rate-limiter-core/requirements.txt
bcrypt==4.3.0
flask==3.1.0
psycopg2-binary==2.9.10
Werkzeug==3.1.3